+++ ChangeLog.rst
@@ -6,6 +6,10 @@ Note worthy changes
 
 - New providers: Exist.io., YNAB, Amazon Cognito.
 
+- You can now store OAuth credentials directly in your
+  ``settings.SOCIALACCOUNT_PROVIDERS`` settings instead of storing them in the
+  database using a ``SocialApp`` record.
+
 
 Backwards incompatible changes
 ------------------------------